Pairwise swap elements of a linked list
Given a singly linked list of size N. The task is to swap elements in the linked list pairwise.
For example, if the input list is 1 2 3 4, the resulting list after swaps will be 2 1 4 3.
Note: You need to swap the nodes, not only the data. If only data is swapped then driver will print -1.

Signed-off-by: Moeez Hayder37 <72216020+Moeez37@users.noreply.github.com>